Problem: json schema needs to be shared accross multiple bins Solution: create json_schema folder
