+++ src/bin/json_schema/mod.rs
@@ -1,0 +1,7 @@
+mod blockchain;
+mod transaction;
+mod machine;
+
+pub use self::blockchain::{JSONVectorBlock, create_block};
+pub use self::transaction::create_transaction;
+pub use self::machine::{create_machine, test_machine};
